for_docker/backend/api/requirements.txt
requests==2.24.0
numpy==1.19.5
Flask==2.0.1
Pillow==8.3.2
torch==1.9.1
